requirement.txt
python-dotenv
streamlit
snowflake-connector-python[secure-local-storage]
snowflake_snowpark_python
IPython
pandas
numpy
dotenv
sseclient-py
streamlit-aggrid
python-docx
pyinstaller
fastapi
uvicorn
python-multipart

# MCP Server Dependencies
mcp>=0.9.0
httpx>=0.25.0

# LangGraph Dependencies (for agentic flow)
langgraph
langchain
langchain-openai
langchain-community
